backend/adk/pyproject.toml
[project]
name = "geminihydra-adk"
version = "1.0.0"
description = "GeminiHydra ADK Orchestration Sidecar"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.2.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "asyncpg>=0.30",
    "httpx>=0.28",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
